+++ src/Test/Scene/webXrTestScene.ts
@@ -139,7 +139,7 @@ export class SceneBuilder implements ISceneBuilder {
 
         mmdRuntime.setManualAnimationDuration(mmdAnimation.endFrame);
 
-        mmdRuntime.setCamera(mmdCamera);
+        mmdRuntime.addAnimatable(mmdCamera);
         mmdCamera.addAnimation(mmdAnimation);
         mmdCamera.setAnimation("motion");